requirements.txt
numpy
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
idna==2.8
lxml==4.4.1
mysql-connector-python==8.0.18
protobuf==3.11.1
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
urllib3==1.25.7
